maya ver 2022 for python3

+++ keysEditFromBuffer/keysEditFromBufferUI.py
@@ -1,6 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
 import sys
 
 from PySide2.QtCore import *
@@ -176,7 +180,7 @@ def main():
 	ptr = OpenMayaUI.MQtUtil.findWindow(mayaWindow)
 	ptr = OpenMayaUI.MQtUtil.mainWindow()
 	if ptr is not None:
-		parent = wrapInstance(long(ptr),QWidget)
+		parent = wrapInstance(int(ptr),QWidget)
 	else:
 		cmds.error('Please open the GraphEditor')
 	app = QApplication.instance()
